requirements.txt
certifi==2019.9.11
geographiclib==1.50
geopy==1.20.0
mysql-connector-python==8.0.18